server/requirements.txt
openenv-core>=0.2.2
fastapi>=0.115.0
uvicorn>=0.24.0
pydantic>=2.0.0
openai>=1.0.0
requests>=2.31.0
